requirements.txt
anthropic>=0.25.0
Pillow>=10.0.0
rich>=13.0.0
click>=8.1.0
requests>=2.31.0
